Add ability to run processes within processes
The Orderable class was created to allow steps and processes to be
treated the same when within another process.

This commit is hardwired to only compile one file with an explicit file
path. A future commit will be made to allow the user to specify the
directory of the file and where Carla should compile to.

Running a process syntax:
run [process name]
after [name], [name]
using ([name expected by process] -> [name passed to the process]),
([name expected by process] -> [name passed to the process])
passing ([name used by the process] -> [name of variable passed]),
([name used by the process] -> [name of variable passed])

Each of the keywords “after”, “using”, and “passing” is optional. It is
possible to use the keywords to specify more than 2 names/pairs; the
commas are only there as an example.