Fix processes trying to run threads they don't own
The writer would write code at the end of every process to start all of
the starting threads from all of the processes. This is because
ScalaWriter was storing the staring threads (Orderables) from a process
in a class variable. This was fine when only one file could be compiled
but broke when multiple processes where compiled by the same static
writer.

The code now has a Map to keep the starting threads separate for each
process. The Map uses the process’ name as a key and the Set of
starting threads (Orderables) as the value.