Remove absolute path to files to include
Files are copied from a .txt version of the Scala files that need to be
included. Text files are not compiled into .jar files so they can be
accessed and copied when the compiler is running. The code could be
improved by finding a way to read the text from the Scala files rather
than duplicating them.